requirements.txt
aioredis==1.3.1
amqp==5.1.1
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.26.5
botocore==1.29.5
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.3
cssselect2==0.7.0
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==1.0.0
Django==4.0.1
django-allauth==0.51.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-formtools==2.4
django-heroku==0.3.1
django-js-asset==2.0.0
django-redis==5.2.0
django-storages==1.13.1
django-timezone-field==5.0
djangorestframework==3.14.0
filelock==3.8.0
future==0.18.2
gunicorn==20.1.0
heroku==0.1.1
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.1
msgpack==1.0.4
oauthlib==3.2.2
oscrypto==1.3.0
packaging==21.3
Pillow==9.3.0
pipenv==2022.10.12
platformdirs==2.5.2
prompt-toolkit==3.0.32
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.3.4
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
smart-open==6.2.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
xhtml2pdf==0.2.8
zope.interface==5.4.0
